Defers to current context and values in ~/.kube/config
We can default to the state machine that is ~/.kube/config quite a bit.
These changes allow us to do so. In lieu of finding any sort of
defaults, we try to create the defaults wherever possible. We stop just
short of putting in the server info as a cluster.